pyproject.toml
[tool.poetry]
name = "starterFlask"
version = "0.1.0"
description = "simple app"
authors = ["majamarijan <majapianist@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]


[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
pillow = "^10.3.0"
flask-caching = "^2.1.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
